requirements.txt
accelerate==1.12.0
addict==2.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
easydict==1.13
einops==0.8.1
filelock==3.20.0
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiwer==4.0.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.2
psutil==7.1.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.20.3
torch==2.9.1+cu130
torchaudio==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
